Move .js test to .ts files
