+++ tests/IntegrationRouter/ModalWithGuard.vue
